Rename navigation menus to dropdown components
Renamed `GalleryNavMenu` to `GalleryNavMenuAsDropdown` and `HeroNavMenu` to `HeroNavMenuAsDropdown` to better reflect their dropdown functionality. Updated all references and imports accordingly.
